refactor(lib): consolidate types in types.ts
Centralises type definitions in types.ts and refactors imports
accordingly
